refactor: handlers and logic code inside viewer component is splitted into different files and constant improvements
